+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added origami client call to delete the active kernel session after successful notebook execution
+
 ## [0.0.11] - 2022-11-18
 ### Added
 - Added API calls to update job instance attempt status during execution